refactoring and restructuring the pyxus library
